// src/app/core/rfid/models/lecture.ts
export interface Lecture {
    IDLectures: number;
    Date: Date;
    IDArticle: number;
    IDChaineMontage: number;
    IDEmploye: number;
    IDOFabrication: number;
    IDLancement: number;
    Quantite: number;
    Temps: number;
    LibTaille: string;
    IDOperation: number;
    NumTicket: number;
    IdPhase: number;
    IDTiers: number;
    IDPacket: number;
    IDGamme: number;
}